build: add RPM and debian packaging
not perfect, and automatic installation of dependencies isn't there yet,
but it's a start.

the idea is to only use cpack's generators (avoid specfiles etc.) as
much as possible.
unfortunately, they're crap.
in particular, in order to create multiple packages (library, headers,
debuginfo), we need to use components, and components aren't
well-supported for unix.

to be specific:
* debuginfo isn't generated.
* depencendices between components aren't generated.
* package names are anything but the conventional names used by redhat
  and debian.

the cpack fix scripts under contrib/package only fix the last issue.

Signed-off-by: Idan Kedar <ik.imbaf@gmail.com>